requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
SQLAlchemy==2.0.31
requests==2.32.3
PyMySQL==1.1.1
python-dotenv==1.0.1
pydantic==2.8.2